Update docs, config arg order, and rename custom_transport module
