+++ config/datadog.spc
@@ -1,11 +1,6 @@
 connection "datadog" {
   plugin = "datadog"
 
-  # The API URL. By default it is pointed to "https://api.datadoghq.com/"
-  # If working with the EU version of Datadog, use "https://api.datadoghq.eu/"
-  # Please note that this URL must not end with the /api/ path.
-  # api_url = "https://api.datadoghq.com/"
-
   # Get your API key from https://app.datadoghq.com/organization-settings/api-keys
   # Steampipe will resolve the API key in below order:
   #   1. The "api_key" specified here in the config
@@ -17,4 +12,9 @@ connection "datadog" {
   #   1. The "app_key" specified here in the config
   #   2. The `DD_CLIENT_APP_KEY` environment variable
   # app_key = "b1cf234c0ed4c567890b524a3b42f1bd91c111a1"
+
+  # The API URL. By default it is pointed to "https://api.datadoghq.com/"
+  # If working with the EU version of Datadog, use "https://api.datadoghq.eu/"
+  # Please note that this URL must not end with the /api/ path.
+  # api_url = "https://api.datadoghq.com/"
 }
