+++ docs/index.md
@@ -60,11 +60,6 @@ Installing the latest datadog plugin will create a config file (`~/.steampipe/co
 connection "datadog" {
   plugin = "datadog"
 
-  # The API URL. By default it is pointed to "https://api.datadoghq.com/"
-  # If working with the EU version of Datadog, use "https://api.datadoghq.eu/"
-  # Please note that this URL must not end with the /api/ path.
-  # api_url = "https://api.datadoghq.com/"
-
   # Get your API key from https://app.datadoghq.com/organization-settings/api-keys
   # Steampipe will resolve the API key in below order:
   #   1. The "api_key" specified here in the config
@@ -76,14 +71,19 @@ connection "datadog" {
   #   1. The "app_key" specified here in the config
   #   2. The `DD_CLIENT_APP_KEY` environment variable
   # app_key = "b1cf234c0ed4c567890b524a3b42f1bd91c111a1"
+
+  # The API URL. By default it is pointed to "https://api.datadoghq.com/"
+  # If working with the EU version of Datadog, use "https://api.datadoghq.eu/"
+  # Please note that this URL must not end with the /api/ path.
+  # api_url = "https://api.datadoghq.com/"
 }
 ```
-
-- `api_url` (required) - The API URL used for all requests. Defaults to "https://api.datadoghq.com/", but should be changed to "https://api.datadoghq.eu/" if working with the EU version.
 
 - `api_key` (required) - [API keys](https://docs.datadoghq.com/account_management/api-app-keys/#api-keys) are unique to an organization. An API key is required by the Datadog Agent to submit metrics and events to Datadog. [Get an API key](https://app.datadoghq.com/organization-settings/api-keys). May alternatively be set via the `DD_CLIENT_API_KEY` environment variable.
 
 - `app_key` (required) - [Application keys](https://docs.datadoghq.com/account_management/api-app-keys/#application-keys) in conjunction with organization’s API key, give users access to Datadog’s programmatic API. Application keys are associated with the user account that created them and have the permissions and capabilities of the user who created them. [Get an application key](https://app.datadoghq.com/organization-settings/application-keys). May alternatively be set via the `DD_CLIENT_APP_KEY` environment variable.
+
+- `api_url` (optional) - The API URL used for all requests. Defaults to "https://api.datadoghq.com/". If working with the EU version, this should be changed to "https://api.datadoghq.eu/".
 
 ## Get Involved
